pav
puv

+++ src/compiler/lib/private/std/types/uint/uint64_lang.rs
@@ -1,16 +1,16 @@
 
 pub struct IntU64Lang{
-    val: u32
+    val: u64
 }
 
 impl IntU64Lang {
-    pub(crate) fn new(val: u32) -> IntU64Lang {
+    pub(crate) fn new(val: u64) -> IntU64Lang {
         IntU64Lang {
             val
         }
     }
 
     fn convert_type_to_c(&self) -> String {
-        "UInt32".to_string()
+        "UInt64".to_string()
     }
 }
